Make graph serialiser the replacement implementation
