+++ Saule/Http/PreprocessingDelegatingHandler.cs
@@ -36,7 +36,6 @@ namespace Saule.Http
         {
             var jsonApi = new JsonApiSerializer();
             jsonApi.JsonConverters.AddRange(config.JsonConverters);
-            jsonApi.UseGraphSerializer = config.UseGraphSerializer;
 
             PrepareQueryContext(jsonApi, request, config);
