+++ Saule/Saule.csproj
@@ -97,11 +97,10 @@
     <Compile Include="Queries\Sorting\SortingContext.cs" />
     <Compile Include="Queries\Sorting\SortingDirection.cs" />
     <Compile Include="Queries\Sorting\SortingProperty.cs" />
-    <Compile Include="Serialization\IResourceSerializer.cs" />
     <Compile Include="Serialization\ResourceGraphPathSet.cs" />
     <Compile Include="Serialization\ResourceGraphRelationship.cs" />
     <Compile Include="Serialization\ResourceGraph.cs" />
-    <Compile Include="Serialization\ResourceGraphSerializer.cs" />
+    <Compile Include="Serialization\ResourceSerializer.cs" />
     <Compile Include="Serialization\ResourceGraphNodeKey.cs" />
     <Compile Include="Serialization\ResourceGraphNode.cs" />
     <Compile Include="ResourceRelationshipOfT.cs" />
@@ -125,7 +124,6 @@
     <Compile Include="Serialization\ErrorSerializer.cs" />
     <Compile Include="Serialization\ResourceDeserializer.cs" />
     <Compile Include="JsonApiException.cs" />
-    <Compile Include="Serialization\ResourceSerializer.cs" />
     <Compile Include="StringExtensions.cs" />
     <Compile Include="TypeExtensions.cs" />
   </ItemGroup>
