+++ Tests/Tests.csproj
@@ -151,7 +151,6 @@
     <Compile Include="Serialization\CanonicalUrlPathBuilderTests.cs" />
     <Compile Include="Serialization\DefaultUrlPathBuilderTests.cs" />
     <Compile Include="Serialization\ResourceGraphNodeKeyTests.cs" />
-    <Compile Include="Serialization\ResourceGraphSerializerTests.cs" />
     <Compile Include="Serialization\ResourceGraphPathSetTests.cs" />
     <Compile Include="Serialization\UrlConstructionTests.cs" />
     <Compile Include="Serialization\ResourceDeserializerTests.cs" />
